FIX: 약속 아이템 간격 설정 ItemDecorater로 변경
