+++ app/src/main/java/com/promi/util/VerticalSpaceItemDecoration.kt
@@ -1,23 +1,21 @@
-package com.promi.view.promise.adapter
+package com.promi.util
 
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class HorizontalSpaceItemDecoration(private val horizontalSpaceWidth: Int) : RecyclerView.ItemDecoration() {
+class VerticalSpaceItemDecoration(private val verticalSpaceHeight: Int) : RecyclerView.ItemDecoration() {
     private var option : Boolean = true // 첫번째 아이템에 대해 간격을 줄것인지
 
     override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
-        // 모든 아이템에 대해 오른쪽에 간격 추가
-        outRect.right = horizontalSpaceWidth
+        outRect.bottom = verticalSpaceHeight
 
         // 첫 번째 아이템에만 왼쪽 간격 추가
         if(option){
             if (parent.getChildAdapterPosition(view) == 0) {
-                outRect.left = horizontalSpaceWidth
+                outRect.top = verticalSpaceHeight
             }
         }
-
     }
 
     fun setOption(option:Boolean){
